Removed website scraper, fixed oculus install dir problem
