--- OculusLibrary/DataExtraction/OculusWebsiteAggregateRating.cs
@@ -1,7 +1,0 @@
-﻿namespace OculusLibrary.DataExtraction
-{
-    public class OculusWebsiteAggregateRating {
-        public int RatingCount { get; set; }
-        public decimal RatingValue { get; set; }
-    }
-}
