--- OculusLibrary/DataExtraction/OculusWebsiteJson.cs
@@ -1,10 +1,0 @@
-﻿namespace OculusLibrary.DataExtraction
-{
-    public class OculusWebsiteJson
-    {
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public OculusWebsiteAggregateRating AggregateRating { get; set; }
-        public string[] Image { get; set; }
-    }
-}
